test(controller): move plugins test to plugins dir
